@@ -1,66 +1,77 @@
 package ru.varasoft.pictureoftheday
 
 import android.content.Context
-import androidx.appcompat.app.AppCompatActivity
+import android.content.Intent
+import android.net.Uri
 import android.os.Bundle
-import android.view.View
-import androidx.constraintlayout.widget.ConstraintLayout
-import androidx.core.content.ContextCompat
-import com.google.android.material.bottomappbar.BottomAppBar
-import com.google.android.material.bottomsheet.BottomSheetBehavior
-import kotlinx.android.synthetic.main.activity_main.*
-import ru.varasoft.pictureoftheday.view.MainFragment
+import androidx.appcompat.app.AppCompatActivity
+import androidx.fragment.app.Fragment
+import geekbarains.material.ui.api.EarthFragment
+import geekbarains.material.ui.api.MarsFragment
+import geekbarains.material.ui.api.WeatherFragment
+import kotlinx.android.synthetic.main.fragment_pod.*
+import ru.varasoft.pictureoftheday.databinding.ActivityMainBinding
+import ru.varasoft.pictureoftheday.view.PODFragment
 import ru.varasoft.pictureoftheday.view.SETTINGS_SHARED_PREFERENCE
+import ru.varasoft.pictureoftheday.view.SettingsFragment
 import ru.varasoft.pictureoftheday.view.THEME_RES_ID
 
-class MainActivity : AppCompatActivity() {
-    var currentTheme: Int? = null
 
-    private lateinit var bottomSheetBehavior: BottomSheetBehavior<ConstraintLayout>
+class MainActivity : AppCompatActivity() {
+    private var _binding: ActivityMainBinding? = null
+    private val binding get() = _binding!!
 
     override fun onCreate(savedInstanceState: Bundle?) {
         val resIdTheme = getSharedPreferences(SETTINGS_SHARED_PREFERENCE, Context.MODE_PRIVATE)
             .getInt(THEME_RES_ID, R.style.Theme_PictureOfTheDay)
         setTheme(resIdTheme)
-        setContentView(R.layout.activity_main)
         super.onCreate(savedInstanceState)
+        _binding = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(binding.root)
         if (savedInstanceState == null) {
-            supportFragmentManager.beginTransaction()
-                .replace(R.id.container, MainFragment.newInstance())
-                .commitNow()
+            replaceFragment(PODFragment.newInstance())
         }
-        setBottomAppBar()
-        //setBottomSheetBehavior(findViewById(R.id.bottom_sheet_container))
+
+        binding.inputLayout.setEndIconOnClickListener {
+            startActivity(Intent(Intent.ACTION_VIEW).apply {
+                data = Uri.parse("https://en.wikipedia.org/wiki/${binding.inputEditText.text.toString()}")
+            })
+        }
+
+        setBottomNavigationView()
     }
 
-    private fun setBottomAppBar() {
-        val context = this
-        context.setSupportActionBar(findViewById(R.id.bottom_app_bar))
-        fab.setOnClickListener {
-            if (isMain) {
-                isMain = false
-                bottom_app_bar.navigationIcon = null
-                bottom_app_bar.fabAlignmentMode = BottomAppBar.FAB_ALIGNMENT_MODE_END
-                fab.setImageDrawable(ContextCompat.getDrawable(context, R.drawable.ic_back_fab))
-                bottom_app_bar.replaceMenu(R.menu.menu_bottom_bar_other_screen)
-            } else {
-                isMain = true
-                bottom_app_bar.navigationIcon =
-                    ContextCompat.getDrawable(context, R.drawable.ic_hamburger_menu_bottom_bar)
-                bottom_app_bar.fabAlignmentMode = BottomAppBar.FAB_ALIGNMENT_MODE_CENTER
-                fab.setImageDrawable(ContextCompat.getDrawable(context, R.drawable.ic_plus_fab))
-                bottom_app_bar.replaceMenu(R.menu.menu_bottom_bar)
+    private fun setBottomNavigationView() {
+        binding.bottomNavigationView.setOnNavigationItemSelectedListener { item ->
+            when (item.itemId) {
+                R.id.bottom_view_telescope -> {
+                    replaceFragment(PODFragment())
+                    true
+                }
+                R.id.bottom_view_earth -> {
+                    replaceFragment(EarthFragment())
+                    true
+                }
+                R.id.bottom_view_mars -> {
+                    replaceFragment(MarsFragment())
+                    true
+                }
+                R.id.bottom_view_weather -> {
+                    replaceFragment(WeatherFragment())
+                    true
+                }
+                R.id.bottom_view_settings -> {
+                    replaceFragment(SettingsFragment())
+                    true
+                }
+                else -> false
             }
         }
     }
 
-    private fun setBottomSheetBehavior(bottomSheet: ConstraintLayout) {
-        bottomSheetBehavior = BottomSheetBehavior.from(bottomSheet)
-        bottomSheetBehavior.state = BottomSheetBehavior.STATE_COLLAPSED
-    }
-
-    companion object {
-        private var isMain = true
-
+    private fun replaceFragment(fragment: Fragment) {
+        supportFragmentManager.beginTransaction()
+            .replace(R.id.container, fragment)
+            .commitNowAllowingStateLoss()
     }
 }
