@@ -1,8 +1,8 @@
-package ru.varasoft.pictureoftheday.model
+package ru.varasoft.pictureoftheday.model.earth
 
 import com.google.gson.annotations.SerializedName
 
-data class PODServerResponseData(
+data class EarthPhotoServerResponseData(
     @field:SerializedName("copyright") val copyright: String?,
     @field:SerializedName("date") val date: String?,
     @field:SerializedName("explanation") val explanation: String?,
@@ -11,5 +11,4 @@ data class PODServerResponseData(
     @field:SerializedName("title") val title: String?,
     @field:SerializedName("url") val url: String?,
     @field:SerializedName("hdurl") val hdurl: String?
-
 )
